refactor: fold the two index-backend crates into quillcache-core (feature-gated modules)
Per request — no need for two separate index crates; merge them into the crate
where MemoryIndex lives (quillcache-core), using file-level (module) separation
instead. The heavy/optional toolchains stay out of the default build via feature
gates, exactly as the separate crates did.

- quillcache-index-holt/src/lib.rs   -> quillcache-core/src/index_holt.rs   (cfg "holt")
- quillcache-index-rocksdb/src/lib.rs -> quillcache-core/src/index_rocksdb.rs (cfg "rocksdb")
  (intra refs rewritten quillcache_core:: -> crate::; only test-mod `use super::*`)
- core Cargo.toml: holt / rocksdb / serde_json become optional deps behind the
  `holt` / `rocksdb` features; default core build still needs neither (rocksdb's
  C++/libclang stays opt-in).
- core re-exports HoltIndex / RocksIndex at the crate root (cfg-gated), so
  quillcache_core::HoltIndex / ::RocksIndex resolve unchanged.
- root binary: `holt` / `rocksdb` features forward to quillcache-core/{holt,rocksdb};
  main.rs + gateway.rs use quillcache_core::{HoltIndex,RocksIndex}; the two
  path deps are gone.
- workspace: dropped both crates from members/default-members (7 -> 5 crates:
  quillcache, -core, -store, -transfer-engine, + excluded -cuda).
- CI: the no-C++ job tests `--workspace` + holt via `-p quillcache-core --features
  holt`; the rocksdb job uses `-p quillcache-core --features rocksdb` (clang).

Verified: default workspace 48 tests; core --features holt 28; core --features
rocksdb 28; CLI builds --features holt,rocksdb; fmt + clippy clean.
